feat: add app configuration tab and user settings
